Adds support for original element's content inside of React
This change request enables a user to access `this.props._content` inside
of the React lifecycle methods. The property will be a node containing any
elements originally located within the web component prior to rendering. This
serves as a close approximation of the <content> tag as used in the template
prior to render.