Improve regexes in parseAttributeValue method
Currently jsonRegexp and jsonArrayRegexp inside parseAttributeValue
contain the same regex (possibly due to c&p), which is wrong. The
testcase involving jsonArrayRegexp does not fail because pointerRegexp
catches the corresponding values. Provide a unique jsonArrayRegexp,
which checks for leading {[ and trailing ]}. Improve pointerRegexp to
check for leading { and trailing }, and jsonRegexp to check for leading
{{ and trailing }}. Also, use replace instead of subtring on the JSON
match and combine the calls to replace on the pointer match.